requirements.txt
Django==4.2
requests==2.28.1
python-dotenv
djangorestframework